requirements.txt
annotated-types==0.7.0asgiref==3.8.1cachetools==5.5.2certifi==2025.4.26chardet==5.2.0charset-normalizer==3.4.2colorama==0.4.6Django==5.2.3django-environ==0.12.0dotenv==0.9.9google-ai-generativelanguage==0.6.15google-api-core==2.25.0google-api-python-client==2.172.0google-auth==2.40.3google-auth-httplib2==0.2.0google-generativeai==0.8.5googleapis-common-protos==1.70.0grpcio==1.73.0grpcio-status==1.71.0httplib2==0.22.0idna==3.10numpypandas==2.3.0proto-plus==1.26.1protobuf==5.29.5psycopg2-binary==2.9.10pyasn1==0.6.1pyasn1_modules==0.4.2pydantic==2.11.5pydantic_core==2.33.2pyparsing==3.2.3python-dateutil==2.9.0.post0python-dotenv==1.1.0pytz==2025.2requests==2.32.4rsa==4.9.1six==1.17.0sqlparse==0.5.3tqdm==4.67.1typing-inspection==0.4.1typing_extensions==4.14.0tzdata==2025.2uritemplate==4.2.0urllib3==2.4.0XlsxWriter==3.2.3